polish ngx RTL support
