+++ generators/client/templates/angular/src/main/webapp/app/layouts/main/_main.component.ts
@@ -16,22 +16,11 @@
  See the License for the specific language governing permissions and
  limitations under the License.
 -%>
-<%_ if (enableI18nRTL) { _%>
-import { Component, OnInit, Renderer } from '@angular/core';
-<%_ } else { _%>
 import { Component, OnInit } from '@angular/core';
-<%_ } _%>
 import { Router, ActivatedRouteSnapshot, NavigationEnd, RoutesRecognized } from '@angular/router';
-<%_ if (enableI18nRTL) {_%>
-import { TranslateService, LangChangeEvent } from '@ngx-translate/core';
-<%_ } _%>
 
 <%_ if (enableTranslation) { _%>
-<%_ if (enableI18nRTL) { _%>
-import { JhiLanguageHelper, StateStorageService, FindLanguageFromKeyPipe } from '../../shared';
-<%_ } else { _%>
 import { JhiLanguageHelper, StateStorageService } from '../../shared';
-<%_ } _%>
 <%_ } else { _%>
 import { Title } from '@angular/platform-browser';
 import { StateStorageService } from '../../shared';
@@ -46,11 +35,6 @@ export class <%=jhiPrefixCapitalized%>MainComponent implements OnInit {
     constructor(
         <%_ if (enableTranslation) { _%>
         private jhiLanguageHelper: JhiLanguageHelper,
-        <%_ if (enableI18nRTL) { _%>
-        private renderer: Renderer,
-        private translateService: TranslateService,
-        private findLanguageFromKeyPipe: FindLanguageFromKeyPipe,
-        <%_ } _%>
         <%_ } else { _%>
         private titleService: Title,
         <%_ } _%>
@@ -65,13 +49,6 @@ export class <%=jhiPrefixCapitalized%>MainComponent implements OnInit {
         }
         return title;
     }
-    <%_ if (enableI18nRTL) { _%>
-
-    private updatePageDirection() {
-        this.renderer.setElementAttribute(document.querySelector('html'), 'lang', this.translateService.currentLang);
-        this.renderer.setElementAttribute(document.querySelector('html'), 'dir', this.findLanguageFromKeyPipe.isRTL(this.translateService.currentLang) ? 'rtl' : 'ltr');
-    }
-    <%_ }_%>
 
     ngOnInit() {
         this.router.events.subscribe((event) => {
@@ -83,10 +60,5 @@ export class <%=jhiPrefixCapitalized%>MainComponent implements OnInit {
                 <%_ } _%>
             }
         });
-        <%_ if (enableI18nRTL) { _%>
-        this.translateService.onLangChange.subscribe((event: LangChangeEvent) => {
-            this.updatePageDirection();
-        });
-        <%_ } _%>
     }
 }
